+++ types/jsdom/package.json
@@ -2,24 +2,19 @@
     "private": true,
     "types": "index",
     "typesVersions": {
-        ">=3.6.0-0": {
+        "<=3.3": {
             "*": [
-                "ts3.6/*"
+                "ts3.3/*"
             ]
         },
-        ">=3.5.0-0": {
-            "*": [
-                "ts3.5/*"
-            ]
-        },
-        ">=3.4.0-0": {
+        "<=3.4": {
             "*": [
                 "ts3.4/*"
             ]
         },
-        ">=3.1.0-0": {
+        "<=3.5": {
             "*": [
-                "ts3.1/*"
+                "ts3.5/*"
             ]
         }
     }
