+++ types/jsdom/ts3.4/index.d.ts
@@ -1,5 +1,5 @@
 // tslint:disable-next-line: no-bad-reference
-/// <reference path="../ts3.1/index.d.ts"/>
+/// <reference path="../ts3.3/index.d.ts"/>
 
 // tslint:disable-next-line: no-declare-current-package no-single-declare-module
 declare module 'jsdom' {
