--- types/jsdom/ts3.6/index.d.ts
@@ -1,10 +1,0 @@
-// tslint:disable-next-line: no-bad-reference
-/// <reference path="../ts3.5/index.d.ts"/>
-
-// tslint:disable-next-line: no-declare-current-package no-single-declare-module
-declare module 'jsdom' {
-    interface DOMWindow {
-        InputEvent: typeof InputEvent;
-        External: typeof External;
-    }
-}
