--- types/jsdom/ts3.6/tsconfig.json
@@ -1,19 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "lib": ["es2015", "dom"],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "strictFunctionTypes": true,
-        "baseUrl": "../../",
-        "typeRoots": ["../../"],
-        "types": [],
-        "noEmit": true,
-        "forceConsistentCasingInFileNames": true
-    },
-    "files": [
-        "jsdom-tests.ts",
-        "index.d.ts"
-    ]
-}
