--- types/jsdom/ts3.6/tslint.json
@@ -1,1 +1,0 @@
-{ "extends": "dtslint/dt.json" }
